% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partial_cm}
\alias{partial_cm}
\title{Compute the global confusion matrix from the FPR and TPR obtained from each node}
\usage{
partial_cm(
  fpr,
  tpr,
  thresholds,
  negative_count,
  total_count,
  descending = FALSE
)
}
\arguments{
\item{fpr}{list - False positive rates for each individual ROC}

\item{tpr}{list - True positive rates for each individual ROC}

\item{thresholds}{list - Thresholds used to compute the fpr and tpr}

\item{negative_count}{list - Total number of samples corresponding to the
negative case}

\item{total_count}{list - Total number of samples}

\item{descending}{thresholds in descending order?}
}
\value{
global confusion matrix and thresholds
}
\description{
Compute the global confusion matrix from the FPR and TPR obtained from each node
}
