% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc-aggregator.R
\name{roc_curve}
\alias{roc_curve}
\title{Compute Receiver operating characteristic (ROC)}
\usage{
roc_curve(fpr, tpr, thresholds, negative_count, total_count)
}
\arguments{
\item{fpr}{list - False positive rates for each individual ROC}

\item{tpr}{list - True positive rates for each individual ROC}

\item{thresholds}{list - Thresholds used to compute the fpr and tpr}

\item{negative_count}{vector - Total number of samples corresponding to the
negative case}

\item{total_count}{vector - Total number of samples}
}
\value{
list with the global fpr, tpr, and thresholds (decreasing)
}
\description{
Compute Receiver operating characteristic (ROC)
}
