% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_vector}
\alias{shift_vector}
\title{Shift a vector left or right according to the value provided}
\usage{
shift_vector(x, n)
}
\arguments{
\item{x}{the vector}

\item{n}{shift}
}
\value{
the vector shifted
}
\description{
Shift a vector left or right according to the value provided
}
\examples{
shift_vector(c(1,2,3,4), 1)
shift_vector(c(1,2,3,4), -1)
}
